requirements.txt
aniso8601==8.0.0
asgiref==3.2.3
certifi==2019.9.11
chardet==3.0.4
Click==7.0
dj-database-url==0.5.0
Django==3.0.2
django-cors-headers==3.2.1
django-google-charts==0.1.1
django-qsstats-magic==1.1.0
djangorestframework==3.11.0
Flask==1.1.1
Flask-RESTful==0.3.7
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
marshmallow==3.2.1
mongoengine==0.18.2
Pillow==6.2.1
psycopg2==2.8.4
pymongo==3.9.0
pyTelegramBotAPI==3.6.6
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
Werkzeug==0.16.0
whitenoise==5.0.1
